Moved the appsync transformer to its own file. Readme updates
