+++ API.md
@@ -54,11 +54,32 @@ new AppSyncTransformer(scope: Construct, id: string, props: AppSyncTransformerPr
 Name | Type | Description 
 -----|------|-------------
 **appsyncAPI**🔹 | <code>[GraphqlApi](#aws-cdk-aws-appsync-graphqlapi)</code> | The cdk GraphqlApi construct.
-**functionResolvers**🔹 | <code>Map<string, Array<[CdkTransformerResolver](#cdk-appsync-transformer-cdktransformerresolver)>></code> | The Lambda Function resolvers designated by the function directive https://github.com/kcwinner/cdk-appsync-transformer#functions.
+**functionResolvers**🔹 | <code>Map<string, Array<[CdkTransformerFunctionResolver](#cdk-appsync-transformer-cdktransformerfunctionresolver)>></code> | The Lambda Function resolvers designated by the function directive https://github.com/kcwinner/cdk-appsync-transformer#functions.
 **nestedAppsyncStack**🔹 | <code>[NestedStack](#aws-cdk-core-nestedstack)</code> | The NestedStack that contains the AppSync resources.
 **outputs**🔹 | <code>[SchemaTransformerOutputs](#cdk-appsync-transformer-schematransformeroutputs)</code> | The outputs from the SchemaTransformer.
 **resolvers**🔹 | <code>any</code> | The AppSync resolvers from the transformer minus any function resolvers.
 **tableNameMap**🔹 | <code>Map<string, any></code> | Map of cdk table tokens to table names.
+
+### Methods
+
+
+#### addLambdaDataSourceAndResolvers(functionName, id, lambdaFunction, options?)🔹 <a id="cdk-appsync-transformer-appsynctransformer-addlambdadatasourceandresolvers"></a>
+
+Adds the function as a lambdaDataSource to the AppSync api Adds all of the functions resolvers to the AppSync api.
+
+```ts
+addLambdaDataSourceAndResolvers(functionName: string, id: string, lambdaFunction: IFunction, options?: DataSourceOptions): LambdaDataSource
+```
+
+* **functionName** (<code>string</code>)  The function name specified in the.
+* **id** (<code>string</code>)  The id to give.
+* **lambdaFunction** (<code>[IFunction](#aws-cdk-aws-lambda-ifunction)</code>)  The lambda function to attach.
+* **options** (<code>[DataSourceOptions](#aws-cdk-aws-appsync-datasourceoptions)</code>)  *No description*
+  * **description** (<code>string</code>)  The description of the data source. __*Default*__: No description
+  * **name** (<code>string</code>)  The name of the data source, overrides the id given by cdk. __*Default*__: generated by cdk given the id
+
+__Returns__:
+* <code>[LambdaDataSource](#aws-cdk-aws-appsync-lambdadatasource)</code>
 
 
 
@@ -168,7 +189,7 @@ Name | Type | Description
 Name | Type | Description 
 -----|------|-------------
 **cdkTables**?🔹 | <code>Map<string, [CdkTransformerTable](#cdk-appsync-transformer-cdktransformertable)></code> | __*Optional*__
-**functionResolvers**?🔹 | <code>Map<string, Array<[CdkTransformerResolver](#cdk-appsync-transformer-cdktransformerresolver)>></code> | __*Optional*__
+**functionResolvers**?🔹 | <code>Map<string, Array<[CdkTransformerFunctionResolver](#cdk-appsync-transformer-cdktransformerfunctionresolver)>></code> | __*Optional*__
 **mutations**?🔹 | <code>Map<string, [CdkTransformerResolver](#cdk-appsync-transformer-cdktransformerresolver)></code> | __*Optional*__
 **noneResolvers**?🔹 | <code>Map<string, [CdkTransformerResolver](#cdk-appsync-transformer-cdktransformerresolver)></code> | __*Optional*__
 **queries**?🔹 | <code>Map<string, string></code> | __*Optional*__
